mark conversation private
