+++ Wire-iOS/Sources/UserInterface/ConversationList/ListContent/ConversationListItemView.swift
@@ -23,7 +23,8 @@ extension Notification.Name {
 
 final class ConversationListItemView: UIView {
     // Please use `updateForConversation:` to set conversation.
-    var conversation: ZMConversation?
+    private var conversation: ZMConversation?
+    
     var titleText: NSAttributedString? {
         didSet {
             titleField.attributedText = titleText
@@ -196,4 +197,84 @@ final class ConversationListItemView: UIView {
         })
     }
 
+
+    func configure(with title: NSAttributedString?, subtitle: NSAttributedString?) {
+        self.titleText = title
+        self.subtitleAttributedText = subtitle
+    }
+    
+    /// configure without a conversation, i.e. when displaying a pending user
+    ///
+    /// - Parameters:
+    ///   - title: title of the cell
+    ///   - subtitle: subtitle of the cell
+    ///   - users: the pending user(s) waiting for self user to accept connection request
+    func configure(with title: NSAttributedString?, subtitle: NSAttributedString?, users: [ZMUser]) {
+        self.titleText = title
+        self.subtitleAttributedText = subtitle
+        self.rightAccessory.icon = .pendingConnection
+        avatarView.configure(context: .connect(users: users))
+        
+        labelsStack.accessibilityLabel = title?.string
+    }
+    
+    func update(for conversation: ZMConversation?) {
+        self.conversation = conversation
+        
+        guard let conversation = conversation else {
+            self.configure(with: nil, subtitle: nil)
+            return
+        }
+        
+        let status = conversation.status
+        
+        // Configure the subtitle
+        var statusComponents: [String] = []
+        let subtitle = status.description(for: conversation)
+        let subtitleString = subtitle.string
+        
+        if !subtitleString.isEmpty {
+            statusComponents.append(subtitleString)
+        }
+        
+        // Configure the title and status
+        let title: NSAttributedString?
+        
+        if ZMUser.selfUser().isTeamMember, let connectedUser = conversation.connectedUser {
+            title = AvailabilityStringBuilder.string(for: connectedUser, with: .list)
+            
+            if connectedUser.availability != .none {
+                statusComponents.append(connectedUser.availability.localizedName)
+            }
+            labelsStack.accessibilityLabel = title?.string
+        } else {
+            title = conversation.displayName.attributedString
+            labelsStack.accessibilityLabel = conversation.displayName
+        }
+        
+        // Configure the avatar
+        avatarView.configure(context: .conversation(conversation: conversation))
+        
+        // Configure the accessory
+        let statusIcon: ConversationStatusIcon?
+        if let player = AppDelegate.shared.mediaPlaybackManager?.activeMediaPlayer,
+            let message = player.sourceMessage,
+            message.conversation == conversation {
+            statusIcon = .playingMedia
+        } else {
+            statusIcon = status.icon(for: conversation)
+        }
+        self.rightAccessory.icon = statusIcon
+        
+        if let statusIconAccessibilityValue = rightAccessory.accessibilityValue {
+            statusComponents.append(statusIconAccessibilityValue)
+        }
+        
+        if conversation.localParticipants.first?.isPendingApproval == true {
+            statusComponents.append("pending approval")
+        }
+        
+        labelsStack.accessibilityValue = FormattedText.list(from: statusComponents)
+        configure(with: title, subtitle: status.description(for: conversation))
+    }
 }
